refactor(utils): consolidate pagination utilities into centralized functions (#112615)
